make s3 access anonymous
