+++ geoglows/data.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from ._constants import get_metadata_table_path, get_sfdc_table_path, get_saber_assign_table_path, get_fdc_table_path, get_sfdc_zarr_uri, get_transformer_table_uri
+from ._constants import get_metadata_table_path, get_sfdc_zarr_uri, get_transformer_table_uri
 from ._download_decorators import _forecast, _retrospective, DEFAULT_REST_ENDPOINT, DEFAULT_REST_ENDPOINT_VERSION
 from .analyze import (
     daily_averages as calc_daily_averages,
@@ -191,6 +191,41 @@ def annual_averages(river_id: int or list, **kwargs) -> pd.DataFrame:
     return calc_annual_averages(df)
 
 
+def sfdc(curve_id: int or list) -> pd.DataFrame:
+    """
+    Retrieves data from the SFDC table based on 'asgn_mid' values for given river_id.
+
+    Args:
+        curve_id (int or list): Single or list of sfdc curve IDs
+
+    Returns:
+        pd.DataFrame
+    """
+    # check that curve_id is a 12 digit integer or a list of such integers
+    if isinstance(curve_id, (int, np.integer)):
+        assert len(str(curve_id)) == 12, "curve_id must be a 12 digit integer"
+    if isinstance(curve_id, list):
+        assert all(len(str(x)) == 12 for x in curve_id), "curve_id must be a 12 digit integer"
+        assert all(isinstance(x, int) for x in curve_id), "curve_id must be a 12 digit integer"
+    ds = xr.open_zarr(get_sfdc_zarr_uri(), storage_options={'anon': True})
+    return ds.sel(curve_id=curve_id).to_dataframe().reset_index()
+
+
+def sfdc_for_river_id(river_id: int or list) -> pd.DataFrame:
+    """
+    Retrieves data from the SFDC table using 'asgn_mid' values obtained from the SABER assign table for the given 'river_id'.
+
+    Args:
+        river_id (int or list): ID(s) of a stream(s).
+
+    Returns:
+        pd.DataFrame: Filtered DataFrame from the SFDC table based on 'asgn_mid' values.
+    """
+    curve_ids = assigned_sfdc_curve_id(river_id)
+    filtered_sfdc = sfdc(curve_ids)
+    return filtered_sfdc
+
+
 def assigned_sfdc_curve_id(river_id: int or list) -> list:
     """
     Retrieves 'asgn_mid' values from the SABER assign table for the given river_id(s).
@@ -217,42 +252,6 @@ def assigned_sfdc_curve_id(river_id: int or list) -> list:
     df = pd.read_parquet(get_transformer_table_uri())
     curve_ids = df.loc[df['river_id'].isin(river_id), 'sfdc_curve_id'].tolist()
     return curve_ids
-
-
-def sfdc(curve_id: int or list) -> pd.DataFrame:
-    """
-    Retrieves data from the SFDC table based on 'asgn_mid' values for given river_id.
-
-    Args:
-        curve_id (int or list): Single or list of sfdc curve IDs
-
-    Returns:
-        pd.DataFrame
-    """
-    # check that curve_id is a 12 digit integer or a list of such integers
-    if isinstance(curve_id, (int, np.integer)):
-        assert len(str(curve_id)) == 12, "curve_id must be a 12 digit integer"
-    if isinstance(curve_id, list):
-        assert all(len(str(x)) == 12 for x in curve_id), "curve_id must be a 12 digit integer"
-        assert all(isinstance(x, int) for x in curve_id), "curve_id must be a 12 digit integer"
-    ds = xr.open_zarr(get_sfdc_zarr_uri())
-    return ds.sel(curve_id=curve_id).to_dataframe().reset_index()
-
-
-def sfdc_for_river_id(river_id: int or list) -> pd.DataFrame:
-    """
-    Retrieves data from the SFDC table using 'asgn_mid' values obtained from the SABER assign table for the given 'river_id'.
-
-    Args:
-        river_id (int or list): ID(s) of a stream(s).
-
-    Returns:
-        pd.DataFrame: Filtered DataFrame from the SFDC table based on 'asgn_mid' values.
-    """
-    curve_ids = assigned_sfdc_curve_id(river_id)
-    filtered_sfdc = sfdc(curve_ids)
-    return filtered_sfdc
-
 
 @_retrospective
 def return_periods(river_id: int or list, *, format: str = 'df', method: str = 'gumbel1') -> pd.DataFrame or xr.Dataset:
@@ -298,37 +297,4 @@ def metadata_tables(columns: list = None, metadata_table_path: str = None) -> pd
     df = pd.read_parquet('http://geoglows-v2.s3-website-us-west-2.amazonaws.com/tables/package-metadata-table.parquet')
     os.makedirs(os.path.dirname(metadata_table_path), exist_ok=True)
     df.to_parquet(metadata_table_path)
-    return df[columns] if columns else df
-
-
-def saber_assign_table(columns: list = None, saber_table_path: str = None) -> pd.DataFrame:
-    """
-    Retrieves the SABER assign table as a pandas DataFrame
-
-    Args:
-        columns (list): optional subset of column names to read from the parquet
-        saber_table_path (str): optional path to a local copy of the SABER assign table
-
-    Returns:
-        pd.DataFrame
-    """
-    if saber_table_path:
-        return pd.read_parquet(saber_table_path, columns=columns)
-    saber_table_path = get_saber_assign_table_path()
-    if os.path.exists(saber_table_path):
-        return pd.read_parquet(saber_table_path, columns=columns)
-
-    warn = f"""
-    Local copy of SABER assign table not found.
-    A copy of the table has been cached at {saber_table_path} which you can move as desired.
-    You should set the environment variable PYGEOGLOWS_SABER_ASSIGN_TABLE_PATH or provide the saber_table_path argument.
-    """
-    warnings.warn(warn)
-
-    # Assuming the S3 URL for the SABER assign table. You may need to adjust this URL.
-    df = pd.read_parquet('http://geoglows-v2.s3-website-us-west-2.amazonaws.com/tables/saber-assign-table.parquet')
-
-    os.makedirs(os.path.dirname(saber_table_path), exist_ok=True)
-    df.to_parquet(saber_table_path)
-
     return df[columns] if columns else df
